:iphone: make home page header more mobile responsive
At a small enough screen width, the "create note" button is now a circle
icon with a plus inside. That's much smaller and more mobile friendly
than a button with the text "create note". Also adjusted padding and
stuff like that to make it prettier. The page is now responsive until
about 230px wide.